Fix minor narrowing conversion warning

+++ src/call.hh
@@ -42,7 +42,7 @@ class alignas(8) CallData {
     Span<uint8_t> old_stack_mem;
     Span<uint8_t> old_heap_mem;
 
-    int used_trampolines = 0;
+    int16_t used_trampolines = 0;
 
     LocalArray<OutArgument, MaxOutParameters> out_arguments;
